adicionando salvamento do codigo quando fecha a janela da ide

+++ main.rs
@@ -11,7 +11,7 @@ println!("{}","aqui vem o for");
 println!("{}","digite um numero");
 let mut num2 = String::new();
 io::stdin().read_line(&mut num2).expect("Falha ao ler a entrada");
-let num2: f64 = num2.trim().parse().expect("Valor inválido");
+let num2: i32 = num2.trim().parse().expect("Valor inválido");
 for y in 2..=5 {
 if y==num2 {
 println!("{}",y);
